+++ src/Hydration/TransformerInterface.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * @copyright   ©2025 Maatify.dev
+ * @Library     Maatify.dev Data Repository
+ * @Project     maatify/data-repository
+ * @author      Mohamed Abdulalim (megyptm) <mohamed@maatify.dev>
+ * @since       2025-11-27 14:00:00
+ * @see         https://www.maatify.dev Maatify.dev
+ * @link        https://github.com/Maatify/data-repository view project on GitHub
+ * @note        Distributed in the hope that it will be useful - WITHOUT WARRANTY.
+ */
+
+namespace Maatify\DataRepository\Hydration;
+
+interface TransformerInterface
+{
+    /**
+     * Transform a value from the data source to the destination property type.
+     *
+     * @param mixed $value
+     * @return mixed
+     */
+    public function transform(mixed $value): mixed;
+}